refactor(gym_scoring): promote to directory, split tests (#1266)
gym_scoring.rs (515 LOC) is promoted to a directory module:
- src/gym_scoring/mod.rs: 219 LOC (down from 515)
- src/gym_scoring/tests.rs: new (~296 LOC)

Pure refactor: zero logic changes, all 19 tests pass.

Closes 18/70 of #1266 (this loop iteration), 27 PRs cumulative since #1285.
